requirements.txt
beautifulsoup4==4.14.3
fitz==0.0.1.dev2
hashids==1.3.1
markdownify==1.2.2
openai==2.24.0
pandas==3.0.1
playwright==1.58.0
pydantic==2.12.5
python_docx==1.2.0
python_pptx==1.0.2
PyYAML==6.0.1
PyYAML==6.0.3
Requests==2.32.5
